refactor(rubric): improve async evaluation and content extraction
- Refactored the `_run_one` function to `_run_single`, simplifying the return type and error handling for asynchronous evaluations.
- Updated the logic to gather results from concurrent tasks, ensuring true parallel execution for multiple runs.
- Moved the `extract_assistant_content` function to the dataset module, enhancing code organization and maintainability.
- Adjusted tests to reflect the new structure and ensure proper functionality of the content extraction method.
